Agent Host changes for sbatten/agents/remove-legacy-reordering-native-dialogs
