+++ src/vs/platform/dialogs/electron-main/dialogMainService.ts
@@ -12,7 +12,8 @@ import { normalizeNFC } from '../../../base/common/normalization.js';
 import { isMacintosh, isWindows } from '../../../base/common/platform.js';
 import { Promises } from '../../../base/node/pfs.js';
 import { localize } from '../../../nls.js';
-import { INativeOpenDialogOptions, massageMessageBoxOptions } from '../common/dialogs.js';
+import { INativeOpenDialogOptions } from '../common/dialogs.js';
+import { massageMessageBoxOptions } from './dialogMainUtils.js';
 import { createDecorator } from '../../instantiation/common/instantiation.js';
 import { ILogService } from '../../log/common/log.js';
 import { IProductService } from '../../product/common/productService.js';
