+++ src/vs/platform/dialogs/test/electron-main/dialogMainUtils.test.ts
@@ -4,9 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { deepEqual } from 'assert';
+import { release } from 'os';
 import { isLinux, isMacintosh, isWindows } from '../../../../base/common/platform.js';
 import { ensureNoDisposablesAreLeakedInTestSuite } from '../../../../base/test/common/utils.js';
-import { IMassagedMessageBoxOptions, massageMessageBoxOptions } from '../../common/dialogs.js';
+import { IMassagedMessageBoxOptions, massageMessageBoxOptions } from '../../electron-main/dialogMainUtils.js';
 import product from '../../../product/common/product.js';
 import { IProductService } from '../../../product/common/productService.js';
 
@@ -83,6 +84,32 @@ suite('Dialog', () => {
 			assertOptions(fourButtonCancel_0, ['2', '3', '4', '1'], 0, 3, [1, 2, 3, 0]);
 			assertOptions(fourButtonCancel_1, ['1', '3', '4', '2'], 0, 3, [0, 2, 3, 1]);
 			assertOptions(fourButtonCancel_2, ['1', '2', '4', '3'], 0, 3, [0, 1, 3, 2]);
+			assertOptions(fourButtonCancel_3, ['1', '2', '3', '4'], 0, 3, [0, 1, 2, 3]);
+			assertOptions(fourButtonCancel_4, ['1', '2', '3', '4'], 0, 4, [0, 1, 2, 3]);
+			assertOptions(fourButtonNegativeCancel, ['1', '2', '3', '4'], 0, -1, [0, 1, 2, 3]);
+		} else if (isMacintosh && Number.parseInt(release(), 10) >= 24) {
+			assertOptions(oneButtonNoCancel, ['1'], 0, 0, [0]);
+			assertOptions(oneButtonCancel_0, ['1'], 0, 0, [0]);
+			assertOptions(oneButtonCancel_1, ['1'], 0, 1, [0]);
+			assertOptions(oneButtonNegativeCancel, ['1'], 0, -1, [0]);
+
+			assertOptions(twoButtonNoCancel, ['1', '2'], 0, 1, [0, 1]);
+			assertOptions(twoButtonCancel_0, ['1', '2'], 0, 0, [0, 1]);
+			assertOptions(twoButtonCancel_1, ['1', '2'], 0, 1, [0, 1]);
+			assertOptions(twoButtonCancel_2, ['1', '2'], 0, 2, [0, 1]);
+			assertOptions(twoButtonNegativeCancel, ['1', '2'], 0, -1, [0, 1]);
+
+			assertOptions(threeButtonNoCancel, ['1', '2', '3'], 0, 2, [0, 1, 2]);
+			assertOptions(threeButtonCancel_0, ['1', '2', '3'], 0, 0, [0, 1, 2]);
+			assertOptions(threeButtonCancel_1, ['1', '2', '3'], 0, 1, [0, 1, 2]);
+			assertOptions(threeButtonCancel_2, ['1', '2', '3'], 0, 2, [0, 1, 2]);
+			assertOptions(threeButtonCancel_3, ['1', '2', '3'], 0, 3, [0, 1, 2]);
+			assertOptions(threeButtonNegativeCancel, ['1', '2', '3'], 0, -1, [0, 1, 2]);
+
+			assertOptions(fourButtonNoCancel, ['1', '2', '3', '4'], 0, 3, [0, 1, 2, 3]);
+			assertOptions(fourButtonCancel_0, ['1', '2', '3', '4'], 0, 0, [0, 1, 2, 3]);
+			assertOptions(fourButtonCancel_1, ['1', '2', '3', '4'], 0, 1, [0, 1, 2, 3]);
+			assertOptions(fourButtonCancel_2, ['1', '2', '3', '4'], 0, 2, [0, 1, 2, 3]);
 			assertOptions(fourButtonCancel_3, ['1', '2', '3', '4'], 0, 3, [0, 1, 2, 3]);
 			assertOptions(fourButtonCancel_4, ['1', '2', '3', '4'], 0, 4, [0, 1, 2, 3]);
 			assertOptions(fourButtonNegativeCancel, ['1', '2', '3', '4'], 0, -1, [0, 1, 2, 3]);
